Bureau is now flexible in width and has a minimum of 7 drawers.
The change in width allows it to compress to suit a smaller screen. This goes against the original goal of having a fixed position to aid muscle-memory, but should be okay as most people don't frequently change screen size.
The MAXLIST change means that clients with fewer than 6 folders will have 7 drawers across the top of the Bureau, but those with more (such as CM) will see all of them. Again, the original intention was to have 7 drawers to aid memory.