Hotfix for TEXTAREA singleUse bug - fixed ERROR, but still flawed.
The code to convert the TEXTAREA into a DIV was flawed and returned
a jQuery object instead of a DOM Node.
Now it produces a readonly TEXTAREA.

The remaining problem is that if you switch Templates, the TEXTAREA
becomes Readonly. The intended fix is to add a "not-saved" attribute to
the PekoeNode (the XML) which is removed when the document is Saved.